+++ packages/client/src/index.ts
@@ -38,6 +38,7 @@ export type {
   PageOptions,
   PageOrientation,
   PageSize,
+  PageSizeKeyword,
   PageSource,
   PageTimings,
   PageWarning,